Extracts invalid_role_error
