+++ app/controllers/concerns/errors.rb
@@ -1,0 +1,10 @@
+module Errors extend ActiveSupport::Concern
+  def invalid_role_error
+    render json: {
+      "errors": [
+        "title": "Role is Invalid",
+        "detail": "User is not allowed to perform this action"
+      ]
+    }, status: 400
+  end
+end
