+++ app/controllers/legacy_tasks_controller.rb
@@ -1,4 +1,6 @@
 class LegacyTasksController < ApplicationController
+  include Errors
+
   before_action :verify_queue_access
   before_action :verify_task_assignment_access, only: [:create, :update]
